Use int type instead of long for fds

+++ benchmarks/src/main/java/org/questdb/LineTCPSenderMainVarLenStrings.java
@@ -52,7 +52,7 @@ public class LineTCPSenderMainVarLenStrings {
         long start = System.nanoTime();
         FilesFacade ff = new FilesFacadeImpl();
         try (Path path = new Path()) {
-            long logFd = -1;
+            int logFd = -1;
             if (args.length == 1) {
                 path.put(args[0]).$();
                 logFd = ff.openRW(path, CairoConfiguration.O_NONE);
@@ -62,19 +62,19 @@ public class LineTCPSenderMainVarLenStrings {
                 for (int i = 0; i < count; i++) {
                     sender.metric("md_msgs");
                     sender
-                            .field("ts_nsec", rnd.nextPositiveLong())
-                            .field("pkt_size", rnd.nextPositiveInt())
-                            .field("pcap_file", nextString(rnd.nextPositiveInt() % 64, rnd))
-                            .field("raw_msg", nextString(rnd.nextPositiveInt() % 512, rnd))
-                            .field("Length", rnd.nextInt())
-                            .field("MsgSeqNum", i)
-                            .field("MsgType", rnd.nextInt() % 1000)
-                            .field("src_ip", rnd.nextString(rnd.nextPositiveInt() % 16))
-                            .field("dst_ip", rnd.nextString(rnd.nextPositiveInt() % 16))
-                            .field("src_port", rnd.nextInt() % 10000)
-                            .field("dst_port", rnd.nextInt() % 10000)
-                            .field("first_dir", rnd.nextBoolean())
-                            .$(i * 10_000_000L);
+                        .field("ts_nsec", rnd.nextPositiveLong())
+                        .field("pkt_size", rnd.nextPositiveInt())
+                        .field("pcap_file", nextString(rnd.nextPositiveInt() % 64, rnd))
+                        .field("raw_msg", nextString(rnd.nextPositiveInt() % 512, rnd))
+                        .field("Length", rnd.nextInt())
+                        .field("MsgSeqNum", i)
+                        .field("MsgType", rnd.nextInt() % 1000)
+                        .field("src_ip", rnd.nextString(rnd.nextPositiveInt() % 16))
+                        .field("dst_ip", rnd.nextString(rnd.nextPositiveInt() % 16))
+                        .field("src_port", rnd.nextInt() % 10000)
+                        .field("dst_port", rnd.nextInt() % 10000)
+                        .field("first_dir", rnd.nextBoolean())
+                        .$(i * 10_000_000L);
                 }
                 sender.flush();
             } finally {
@@ -97,11 +97,11 @@ public class LineTCPSenderMainVarLenStrings {
 
     private static class LoggingLineChannel implements LineChannel {
         private final FilesFacade ff;
-        private final long outFileFd;
+        private final int outFileFd;
         private LineChannel delegate;
         private long fileOffset = 0;
 
-        private LoggingLineChannel(LineChannel delegate, long outFileFd, FilesFacade ff) {
+        private LoggingLineChannel(LineChannel delegate, int outFileFd, FilesFacade ff) {
             this.delegate = delegate;
             this.outFileFd = outFileFd;
             this.ff = ff;
